Modified GUI to allow user control of whether tables are spatially indexed and de-duplicated.
Fixed #5

+++ QGIS Plugin/indexer_thread.py
@@ -29,27 +29,31 @@ class IndexerThread(QThread):
     error = pyqtSignal(str)
     progressChanged = pyqtSignal(int)
     
-    def __init__(self, cur, schema, tables):
+    def __init__(self, cur, schema, tables, createSpatialIndex=True, dedup=True):
         QThread.__init__(self)
         self.cur = cur
         self.schema = schema
         self.tables = tables
+        self.createSpatialIndex = createSpatialIndex
+        self.dedup = dedup
 
     def run(self):
+        
         # Todo parallelise this function
         qDic = {}
         i = 0
         for table in self.tables:
-            try:
-                self.cur.execute("""CREATE INDEX """ + table + """_wkb_geometry_gist ON """ + self.schema + """.""" + table + """ USING gist (wkb_geometry)""", qDic)
-            except psycopg2.ProgrammingError, e:
-                if e.message.strip().startswith('relation') and e.message.strip().endswith('does not exist') or \
-                   e.message.strip() == 'column "wkb_geometry" does not exist':
-                    # There were no matching features imported so the table was not created, do not index
-                    pass
-            else:
-                if self.cur.statusmessage != 'CREATE INDEX':
-                    error.emit('Failed to create index on geometry for %s.%s' % (self.schema, table))
+            if self.createSpatialIndex:
+                try:
+                    self.cur.execute("""CREATE INDEX """ + table + """_wkb_geometry_gist ON """ + self.schema + """.""" + table + """ USING gist (wkb_geometry)""", qDic)
+                except psycopg2.ProgrammingError, e:
+                    if e.message.strip().startswith('relation') and e.message.strip().endswith('does not exist') or \
+                       e.message.strip() == 'column "wkb_geometry" does not exist':
+                        # There were no matching features imported so the table was not created, do not index
+                        pass
+                else:
+                    if self.cur.statusmessage != 'CREATE INDEX':
+                        error.emit('Failed to create index on geometry for %s.%s' % (self.schema, table))
             i += 1
             progress = int( float(i) / (len(self.tables)*3) * 100.0 )
             self.progressChanged.emit(progress)
@@ -70,24 +74,33 @@ class IndexerThread(QThread):
             progress = int( float(i) / (len(self.tables)*3) * 100.0 )
             self.progressChanged.emit(progress)
             
+            if self.dedup:
+                try:
+                    # Create a quick hash index
+                    self.cur.execute("""CREATE INDEX """ + table + """_fid_hash ON """ + self.schema + """.""" + table + """ USING hash (fid)""", qDic)
+                    # De-dup the table
+                    self.cur.execute("""DELETE FROM 
+                                            """ + self.schema + """.""" + table + """ 
+                                        USING 
+                                            """ + self.schema + """.""" + table + """ AS dup
+                                        WHERE 
+                                            """ + self.schema + """.""" + table + """.fid = dup.fid AND 
+                                            """ + self.schema + """.""" + table + """.ogc_fid < dup.ogc_fid;""", qDic)
+                    # Drop the index
+                    self.cur.execute("""DROP INDEX """ + table + """_fid_hash""", qDic)
+                except psycopg2.ProgrammingError, e:
+                    if e.message.strip().startswith('relation') and e.message.strip().endswith('does not exist'):
+                        # There were no matching features imported so the table was not created, do not index
+                        pass
             try:
-                # Create a quick hash index
-                self.cur.execute("""CREATE INDEX """ + table + """_fid_hash ON """ + self.schema + """.""" + table + """ USING hash (fid)""", qDic)
-                # De-dup the table
-                self.cur.execute("""DELETE FROM 
-                                        """ + self.schema + """.""" + table + """ 
-                                    USING 
-                                        """ + self.schema + """.""" + table + """ AS dup
-                                    WHERE 
-                                        """ + self.schema + """.""" + table + """.fid = dup.fid AND 
-                                        """ + self.schema + """.""" + table + """.ogc_fid < dup.ogc_fid;""", qDic)
-                # Drop the index
-                self.cur.execute("""DROP INDEX """ + table + """_fid_hash""", qDic)
+                self.cur.execute("""ANALYZE """ + self.schema + """.""" + table, qDic)
             except psycopg2.ProgrammingError, e:
                 if e.message.strip().startswith('relation') and e.message.strip().endswith('does not exist'):
-                    # There were no matching features imported so the table was not created, do not index
+                    # There were no matching features imported so the table was not created, do not analyse
                     pass
-            self.cur.execute("""ANALYZE """ + self.schema + """.""" + table, qDic)
+            else:
+                if not self.cur.statusmessage.startswith('ANALYZE'):
+                    error.emit('Failed to analyze %s.%s' % (self.schema, table))
             i += 1
             progress = int( float(i) / (len(self.tables)*3) * 100.0 )
             self.progressChanged.emit(progress)
